requirements.txt
Django==2.2
pytz==2019.3
sqlparse==0.3.0
django-phone-field==1.8.0
django-widget-tweaks==1.4.5
parse==1.14.0
parse-type==0.5.2
Pillow==7.0.0
selenium==3.141.0
six==1.13.0
urllib3==1.25.7